pyproject.toml
[tool.poetry]
name = "check-backend"
version = "0.1.0"
description = "Minimal FastAPI backend with PostgreSQL and SQLAlchemy"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = { extras = ["standard"], version = "^0.23.0" }
sqlalchemy = "^2.0.20"
psycopg = {extras = ["binary"], version = "^3.3.3"}
bcrypt = "^4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
pydantic-settings = "^2.13.1"
pydantic = {extras = ["email"], version = "^2.13.1"}
minio = "^7.2.7"
python-multipart = "^0.0.26"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
